promoted common functionality to super
